DN-3667 (refactor) moved helper classes to .Helpers namespace
