+++ UvA.Workflow.Tests/Controllers/Helpers/ControllerTestsHelpers.cs
@@ -2,7 +2,7 @@ using MongoDB.Bson;
 using UvA.Workflow.Entities.Domain;
 using UvA.Workflow.Users;
 
-namespace UvA.Workflow.Tests.Controllers;
+namespace UvA.Workflow.Tests.Controllers.Helpers;
 
 /// <summary>
 /// Shared constants and factory methods for impersonation tests.
