+++ UvA.Workflow.Tests/Controllers/ScreensControllerTests.cs
@@ -2,6 +2,7 @@ using Microsoft.AspNetCore.Mvc;
 using Moq;
 using UvA.Workflow.Api.Screens;
 using UvA.Workflow.Api.Screens.Dtos;
+using UvA.Workflow.Tests.Controllers.Helpers;
 
 namespace UvA.Workflow.Tests.Controllers;
 
@@ -11,9 +12,9 @@ public class ScreensControllerTests : ControllerTestsBase
 
     public ScreensControllerTests() : base()
     {
-        _screenDataService = new ScreenDataService(_modelService, _instanceService, _instanceRepoMock.Object,
+        _screenDataService = new ScreenDataService(_modelService, _instanceService, _workflowInstanceRepoMock.Object,
             new InstanceAuthorizationFilterService(_rightsService, _modelService, _userServiceMock.Object,
-                _instanceRepoMock.Object));
+                _workflowInstanceRepoMock.Object));
     }
 
     [Theory]
